# [GLUTEN-4243][CH] Improve sparkFloor function performance with avx2