## What changes were proposed in this pull request?

(Please fill in changes proposed in this fix)   

Fixes: \#4243    cc @KevinyhZou 


``` bash
./build_gcc/utils/extern-local-engine/tests/benchmark_local_engine --benchmark_filter="BM.*Floor.*Float64"               


# 优化前
----------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
BM_CHFloorFunction_For_Float64         36547 ns        36547 ns        18987
BM_SparkFloorFunction_For_Float64     134542 ns       134537 ns         5219



# 优化1：去掉checkAndSetNullable中的if, 自动simd，相比优化前提升10%
----------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
BM_CHFloorFunction_For_Float64         36972 ns        36971 ns        18573
BM_SparkFloorFunction_For_Float64     121571 ns       121566 ns         5756


# 优化2: intel intrinsics手动simd，相比优化前提升14%
----------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
BM_CHFloorFunction_For_Float64         36880 ns        36878 ns        18867
BM_SparkFloorFunction_For_Float64     118141 ns       118135 ns         5939 

```