@taiyang-li 2024-01-05
Update benchmark report after changes of pr: https://github.com/oap-project/gluten/pull/4250/files



``` bash
./build_gcc/utils/extern-local-engine/tests/benchmark_local_engine --benchmark_filter="BM.*Floor.*Float64"                 


# 自动simd
----------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
BM_CHFloorFunction_For_Float64         36268 ns        36266 ns        19117
BM_SparkFloorFunction_For_Float64     112999 ns       112994 ns         6158

# 手动simd
----------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
BM_CHFloorFunction_For_Float64         35942 ns        35940 ns        18993
BM_SparkFloorFunction_For_Float64     108976 ns       108972 ns         6416
```